analysis/requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy
pdfplumber
python-docx
pandas
openpyxl
Pillow
pytesseract
anthropic
python-dotenv
